requirements.txt
aiofiles==25.1.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
bcrypt==5.0.0
beautifulsoup4==4.14.3
certifi==2026.2.25
charset-normalizer==3.4.5
click==8.3.1
colorama==0.4.6
Deprecated==1.3.1
dnspython==2.8.0
ecdsa==0.19.1
fastapi==0.135.1
filelock==3.25.2
greenlet==3.3.2
h11==0.16.0
idna==3.11
joblib==1.5.3
limits==5.8.0
lxml==6.0.2
numpy==2.4.3
nvidia-nccl-cu12==2.29.7
packaging==26.0
pandas==3.0.1
passlib==1.7.4
psycopg2-binary==2.9.11
pyasn1==0.6.2
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-jose==3.5.0
python-whois==0.9.6
requests==2.32.5
requests-file==3.0.1
rsa==4.9.1
scikit-learn==1.8.0
scipy==1.17.1
six==1.17.0
slowapi==0.1.9
soupsieve==2.8.3
SQLAlchemy==2.0.48
starlette==0.52.1
threadpoolctl==3.6.0
tldextract==5.3.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.41.0
wrapt==2.1.2
xgboost==3.2.0
